fix: fallback to pre-#27 code on darwin without cgo
keeping the old shell way as fallback in case of API error.

This change ensures the fallback is used also when CGO is not available
compile time.

Github workflow is updated to add a step where tests are run without CGO.

Testing Done: Github actions and locally:  go test ./... with CGO_ENABLED=0 and 1 on MacOS 14 VM